Updated the look of it
Glitches: when adding a picture after the item was created, the item still claims to not have a picture.
When we view an already taken picture, we should send it to a different screen without the confirm button.
We could update to have a cleaner version of frequency so that the user can choose how long the frequency will go on for and maybe display them only one at a time so that the entire app doesn't get overwhelmed