Complete issue #9
In order to make frequency page less busy, I made it to where only the most urgent occurrence of an item (the current due or the past due) will appear. The rest of the occurrences will not exist until those are completed. Completed functionality was not implemented, so I added that to the app as well. It was specified in the issue that there was a desire for it to be less cluttered by completing items, so I feel this is within the issue still. I made it so there is a completed filter for completed items to keep the list view less busy. I also changed the colors of the bill circles to match the status of the bill. Only creating the items when they were the next one due made the calendar view page a bit wonky. To fix this, I created a function called getFutureOccurrences() that would display the dot on the calendar to show an item is due that day without creating the next year's worth of bills.